lib/utils.py
import numpy as np
import torch


device = torch.device('cuda:0') if torch.cuda.is_available() else torch.device('cpu')


def get_local_device() -> torch.device:
    return device


def ohe_labels(y: torch.Tensor, classes_cnt: int) -> torch.Tensor:
    batch_size = y.shape[0]
    m = torch.zeros(batch_size, classes_cnt).to(y.device)
    m[np.arange(batch_size), y.long()] = 1
    return m


def calc_grad_norm(model) -> float:
    total_norm = 0.
    for p in model.parameters():
        param_norm = p.grad.data.norm(2)
        total_norm += param_norm.item() ** 2
    total_norm = total_norm ** (1. / 2)
    return total_norm
