requirements.txt
pytorch==2.1
torchvision
requests
matplotlib
wandb
tqdm
